fix(sdf-stats): use 64 bit counters for the sphere tracing statistics
The march totals and the histogram sum only stay inside 32 bits for short renders: an 800x800/3000spp render produces 5.3e9 marches, so the accumulator wrapped and the printed average (368 steps per march instead of 69) looked merely odd rather than broken. Both the accumulator and the histogram buckets are now unsigned long long. Verified image-neutral: rebuilding with the 64 bit counters reproduces the 400x400/500spp render bit for bit (max diff 0) and the same statistics.

+++ src/intersections.h
@@ -94,7 +94,8 @@ __host__ __device__ float sphereIntersectionTest(
  *
  * @param stepCounter  Optional instrumentation: adds the number of marching
  *                     steps this call used (may be NULL).
- * @param histogram    Optional instrumentation: 16 buckets of 8 steps (may be
+ * @param histogram    Optional instrumentation: 16 buckets of 8 steps, as 64 bit
+ *                     counters because a long render overflows 32 of them (may be
  *                     NULL).
  * @return             Ray parameter `t` value. -1 if no intersection.
  */
@@ -105,7 +106,7 @@ __host__ __device__ float sdfIntersectionTest(
     glm::vec3& normal,
     bool& outside,
     unsigned long long* stepCounter,
-    unsigned int* histogram);
+    unsigned long long* histogram);
 
 /**
  * Evaluate the signed distance field of `geom` at an object space point. Units:
